requirements.txt
Flask==2.3.3
openai>=0.27.8
pandas>=1.3.0
numpy>=1.21.0
azure-storage-blob==12.17.0
Jinja2==3.1.2
python-dotenv==0.21.1
pypdf>=3.11.1
scikit-learn>=1.3.1
scipy<=1.11.1
plotly>=5.17.0
matplotlib==3.8.0
pymupdf==1.22.5

# Flask==1.1.2
# openai>=0.27.8
# pandas>=1.3.0
# numpy>=1.21.0
# scikit-learn
# scipy<=1.11.1
# pypdf>=3.11.1
# langchain>=0.0.220
# gunicorn==20.0.4
# Jinja2==2.11.3
# MarkupSafe==1.1.1
# Werkzeug==1.0.1
# itsdangerous==1.1.0
# matplotlib==3.7.1
# click==7.1.2
# plotly>=5.0.
# pymupdf==1.22.5
# azure-storage-blob==12.17.0
# python-dotenv==0.21.1


